Refactor updates to the IDR static website
Similarly to a change that was made for www.openmicroscopy.org, this
removes the requirement to track the explicit version of the static
website deployed at https://idr.openmicroscopy.org/about/.
The ome.deploy_archive job is replaced by a simple Python script which
polls hourly for new GitHub releases and deploys them if it finds them.

The environment version is stored as a source of truth under /srv/www/VERSION
and copied as part of the deployment script.