Consume version 0.3.6 of the OMERO Prometheus role
OMERO.server deployed using the latest release of omero-certificates
no longer supports the deprecated TLS 1.0 and 1.1 protocols.
The monitoring agent needs to create an OMERO.py client able to
negotiate a connection with TLS 1.2. The latest version of the role
uncaps the version of OMERO.py installed in the virtual environment
which was previously set to 5.6.0.